requirements.txt
click==8.1.3
Flask==2.2.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.30
protobuf==3.20.1
SQLAlchemy==1.4.40
Werkzeug==2.2.1
zipp==3.8.1
